Organização das rotas, Alteração no layout de login
