+++ views/login.ejs
@@ -3,42 +3,19 @@
 
 <head>
 	<%- include("."+path+"/partials/head") %>
+	<link rel="stylesheet" href="/content/login.css">
 </head>
 <body>
-
-	<!-- Navigation -->
-	<nav class="navbar navbar-inverse navbar-fixed-top" role="navigation">
-		<div class="container">
-			<%- include("."+path+"/partials/menu") %>
-		</div>
-		<!-- /.container -->
-	</nav>
-
-
-	<div class="container">
-		<div class="col-sm-6 col-sm-offset-3">
-
-			<h1><span class="fa fa-sign-in"></span> Login</h1>
-
-			<% if (message.length > 0) { %>
-				<div class="alert alert-danger"><%= message %></div>
-			<% } %>
-
-			<!-- LOGIN FORM -->
-			<form action="./login" method="post">
-				<div class="form-group">
-					<label>User</label>
-					<input type="text" class="form-control" name="username">
-				</div>
-				<div class="form-group">
-					<label>Password</label>
-					<input type="password" class="form-control" name="password">
-				</div>
-
-				<button type="submit" class="btn btn-warning btn-lg" style="float: right;">Sign in</button>
-			</form>
-
-		</div>
+	<div class="login">
+		<h1>Login</h1>
+		<form action="./login" method="post">
+			<input type="text" name="username" placeholder="Username" required="required" />
+			<input type="password" name="password" placeholder="Password" required="required" />
+			<button type="submit" class="btn btn-primary btn-block btn-large">Sign in</button>
+		</form>
+		<% if (message.length > 0) { %>
+			<div class="alert alert-danger" style="margin-top: 12px;"><%= message %></div>
+		<% } %>
 	</div>
 </body>
 </html>
